chore: Redoe everything :sparkles:

+++ augmented-shop/augmented-shop-bukkit/src/main/kotlin/dev/slne/augmented/shop/bukkit/listeners/shop/ShopDestroyListener.kt
@@ -1,10 +1,13 @@
 package dev.slne.augmented.shop.bukkit.listeners.shop
 
 import com.github.shynixn.mccoroutine.folia.entityDispatcher
-import dev.slne.augmented.common.base.bukkit.plugin.plugin
+import com.github.shynixn.mccoroutine.folia.launch
+import dev.slne.augmented.common.base.bukkit.extensions.toPosition
+import dev.slne.augmented.shop.api.shopManager
 import dev.slne.augmented.shop.bukkit.extensions.getShop
+import dev.slne.augmented.shop.bukkit.plugin
 import dev.slne.augmented.shop.core.CoreShop
-import kotlinx.coroutines.withContext
+import dev.slne.augmented.shop.core.CoreShopManager
 import net.kyori.adventure.sound.Sound
 import net.kyori.adventure.text.Component
 import net.kyori.adventure.text.format.NamedTextColor
@@ -15,12 +18,19 @@ import org.bukkit.event.block.BlockBreakEvent
 object ShopDestroyListener : Listener {
 
     @EventHandler
-    suspend fun BlockBreakEvent.onBreak() {
+    fun BlockBreakEvent.onBreak() {
         val shop = block.getShop() as CoreShop? ?: return
+        val shopManager = shopManager as CoreShopManager
 
-        shop.delete()
+        if (shopManager.isLocationLocked(block.location.world.uid, block.location.toPosition())) {
+            player.sendMessage("locked break")
+            isCancelled = true
+            return
+        }
 
-        withContext(plugin.entityDispatcher(player)) {
+        plugin.launch(plugin.entityDispatcher(player)) {
+            shop.delete()
+
             player.playSound(
                 Sound.sound().type(org.bukkit.Sound.BLOCK_ANVIL_BREAK).build(),
                 Sound.Emitter.self()
